소수점이 무한대로 나오거나 서로 다른 식단명의 길이 차이로 인해, 레이아웃이 밀리거나 글자가 겹치는 이슈 해결

+++ AIFoodApp/app/src/main/java/com/android/aifoodapp/RecyclerView/FoodInfoAdapter.java
@@ -52,7 +52,7 @@ public class FoodInfoAdapter extends RecyclerView.Adapter<FoodInfoAdapter.Custom
         /* setting */
         holder.cl_foodInfo.setImageResource(Integer.parseInt(items.get(position).getCl_img()));
         holder.cl_foodName.setText(items.get(position).getFood().getName());
-        holder.cl_caloriesInfo.setText(String.valueOf(items.get(position).getFood().getCalorie()));
+        holder.cl_caloriesInfo.setText(String.valueOf((int) items.get(position).getFood().getCalorie()));
         holder.cl_intake.setText(String.valueOf(items.get(position).getCl_intake()));
         holder.food_list_name.setText(items.get(position).getFood().getName());
